conferenceApp/requirements.txt
channels
Django
django-cors-headers
django-rest-framework-social-oauth2
djangorestframework
drf-yasg
oauthlib
packaging
psycopg2-binary
python-decouple
social-auth-app-django


